+++ glue/app/qt/application.py
@@ -14,7 +14,8 @@ from glue.core import command, Data
 from glue import env
 from glue.main import load_plugins
 from glue.qt import get_qapp
-from glue.qt.qtutil import data_wizard, load_ui, get_icon, action
+from glue.qt.qtutil import load_ui, get_icon, action
+from glue.dialogs.data_wizard.qt import data_wizard
 from glue.app.qt.edit_subset_mode_toolbar import EditSubsetModeToolBar
 from glue.app.qt.mdi_area import GlueMdiArea, GlueMdiSubWindow
 from glue.app.qt.layer_tree_widget import PlotAction, LayerTreeWidget
