+++ glue/dialogs/data_wizard/qt/__init__.py
@@ -1,0 +1,1 @@
+from .data_wizard_dialog import *
